Make peace with no cleanup
